requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.2
django-phonenumber-field==7.0.2
django-request-logging==0.7.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
flake8==6.0.0
isort==5.12.0
mccabe==0.7.0
phonenumbers==8.13.7
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.7.1
